+++ src/main/kotlin/com/flyn/sarcopenia_server/gui/UserFile.kt
@@ -1,0 +1,5 @@
+package com.flyn.sarcopenia_server.gui
+
+import java.util.*
+
+data class UserFile(val uuid: UUID, val fileName: String)