+++ src/main/kotlin/com/flyn/sarcopenia_server/handler/FileMsgHandler.kt
@@ -1,6 +1,8 @@
 package com.flyn.sarcopenia_server.handler
 
 import com.flyn.fc_message.message.FileMessage
+import com.flyn.sarcopenia_server.FileManager
+import com.flyn.sarcopenia_server.gui.UserFile
 import io.netty.channel.ChannelHandlerContext
 import io.netty.channel.SimpleChannelInboundHandler
 import java.io.File
@@ -10,18 +12,13 @@ import java.nio.channels.FileChannel
 
 class FileMsgHandler: SimpleChannelInboundHandler<FileMessage>() {
 
-    companion object {
-        private const val storagePath = "C:/Users/etern/Desktop/receiveData"
-//        private const val storagePath = "./receiveData"
-    }
-
     private var fileChannels: FileChannel? = null
 
     override fun channelRead0(ctx: ChannelHandlerContext, msg: FileMessage) {
         if (fileChannels == null) {
             try {
-                val path = "$storagePath/${msg.uuid}/${msg.fileName.replace(":", "_")}"
-                File("$storagePath/${msg.uuid}").let { dir ->
+                val path = "${FileManager.STORAGE_PATH}/${msg.uuid}/${msg.fileName.replace(":", "_")}"
+                File("${FileManager.STORAGE_PATH}/${msg.uuid}").let { dir ->
                     if (!dir.exists()) dir.mkdirs()
                 }
                 File(path).createNewFile()
@@ -38,6 +35,11 @@ class FileMsgHandler: SimpleChannelInboundHandler<FileMessage>() {
                 println("file receive finish")
                 channel.close()
                 fileChannels = null
+                UserFile(msg.uuid, msg.fileName).takeIf {
+                    !FileManager.FILE_LIST.contains(it)
+                }?.let {
+                    FileManager.FILE_LIST.add(it)
+                }
             }
         }?: run { println("not find channel") }
     }