+++ CHANGELOG.md
@@ -5,8 +5,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- GUI with server control function and file viewer
 
 ## [0.1.0] - 2022-03-18
 ### Added
@@ -33,4 +31,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.0] - 2022-10-12
 ### Added
-- GUI with compose desktop library
+- GUI with server control function and file viewer